First version of Criptoquip crack
This is a script written in python, still not polished, to crack a password using frequency analysis and regular expressions, and obviously the magic of Python. You only need to run the script and have an interpreter configured to display the results. You NEED to have the two text files to start analyzing, if not, program will not run.